--- modules/apps/portal-settings/portal-settings-web/src/main/java/com/liferay/portal/settings/web/internal/portal/settings/configuration/admin/display/LanguageEditCompanyPortalSettingsConfigurationScreenWrapper.java
@@ -1,71 +1,0 @@
-/**
- * SPDX-FileCopyrightText: (c) 2023 Liferay, Inc. https://liferay.com
- * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
- */
-
-package com.liferay.portal.settings.web.internal.portal.settings.configuration.admin.display;
-
-import com.liferay.configuration.admin.display.ConfigurationScreen;
-import com.liferay.configuration.admin.display.ConfigurationScreenWrapper;
-import com.liferay.portal.settings.configuration.admin.display.PortalSettingsConfigurationScreenContributor;
-import com.liferay.portal.settings.configuration.admin.display.PortalSettingsConfigurationScreenFactory;
-
-import javax.servlet.ServletContext;
-
-import org.osgi.service.component.annotations.Component;
-import org.osgi.service.component.annotations.Reference;
-
-/**
- * @author Drew Brokke
- */
-@Component(service = ConfigurationScreen.class)
-public class LanguageEditCompanyPortalSettingsConfigurationScreenWrapper
-	extends ConfigurationScreenWrapper {
-
-	@Override
-	protected ConfigurationScreen getConfigurationScreen() {
-		return _portalSettingsConfigurationScreenFactory.create(
-			new LanguageEditCompanyPortalSettingsConfigurationScreenContributor());
-	}
-
-	@Reference
-	private PortalSettingsConfigurationScreenFactory
-		_portalSettingsConfigurationScreenFactory;
-
-	@Reference(
-		target = "(osgi.web.symbolicname=com.liferay.portal.settings.web)"
-	)
-	private ServletContext _servletContext;
-
-	private class
-		LanguageEditCompanyPortalSettingsConfigurationScreenContributor
-			implements PortalSettingsConfigurationScreenContributor {
-
-		@Override
-		public String getCategoryKey() {
-			return "localization";
-		}
-
-		@Override
-		public String getJspPath() {
-			return "/language.jsp";
-		}
-
-		@Override
-		public String getKey() {
-			return "language";
-		}
-
-		@Override
-		public String getSaveMVCActionCommandName() {
-			return "/portal_settings/edit_company";
-		}
-
-		@Override
-		public ServletContext getServletContext() {
-			return _servletContext;
-		}
-
-	}
-
-}
